Try to collissions

+++ src/com/tooflya/bouncekid/entity/Personage.java
@@ -47,7 +47,7 @@ public class Personage extends Entity {
 		this.setIgnoreUpdate(false);
 	}
 
-	public Personage(final int x, final int y, final TiledTextureRegion pTiledTextureRegion) {
+	public Personage(final float x, final float  y, final TiledTextureRegion pTiledTextureRegion) {
 		this(pTiledTextureRegion);
 
 		this.setPosition(x, y);
